# src/passenger/src/docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PassengerType": {
				"enum": [
					0,
					1,
					2,
					3
				],
				"type": "number"
			},
			"PassengerDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"passportNumber": {
						"type": "string"
					},
					"passportType": {
						"$ref": "#/components/schemas/PassengerType"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"age",
					"passportNumber",
					"passportType",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "passenger",
		"version": "1.0.0",
		"description": "passenger service",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Meysam Hadeli"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/passenger/v1/get-by-id": {
			"get": {
				"operationId": "GetPassengerById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PassengerDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/passenger/v1/get-all": {
			"get": {
				"operationId": "GetPassengers",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PassengerDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"default": "ASC",
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "id",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "searchTerm",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}